First impl, test required

+++ cmd/metal-api/internal/service/v1/machine.go
@@ -55,7 +55,7 @@ type MachineNetwork struct {
 	Prefixes            []string `json:"prefixes" description:"the prefixes of this network"`
 	IPs                 []string `json:"ips" description:"the ip addresses of the allocated machine in this vrf"`
 	Vrf                 uint     `json:"vrf" description:"the vrf of the allocated machine"`
-	ASN                 int64    `json:"asn" description:"ASN number for this network in the bgp configuration"`
+	ASN                 uint32   `json:"asn" description:"ASN number for this network in the bgp configuration"`
 	Private             bool     `json:"private" description:"indicates whether this network is the private network of this machine"`
 	Nat                 bool     `json:"nat" description:"if set to true, packets leaving this network get masqueraded behind interface ip"`
 	DestinationPrefixes []string `json:"destinationprefixes" modelDescription:"prefixes that are reachable within this network" description:"the destination prefixes of this network"`
